* Manafinder: Remove requirement to select order to see data * Implement Leaderboard * Update links on homepage
