+++ styles/app.css
@@ -5,4 +5,9 @@
 html,
 body {
   font-family: "EB Garamond", serif, sans-serif;
+  height: 100%;
+}
+
+#__next {
+  height: 100%;
 }
